Auslagerung der FileOperation im deutschen Pad
